mul_along_dim pushthrough
When this goes well, it reduces the total number of multiplies that the
GPU has to do, and it creates opportunities to merge operations like
sum_multi and shuffle to vanish.

Some of this is hidden behind an off-by-default flag, because the
pushthrough is currently hard to control and sometimes pushes too far,
making changes that aren't advantageous.